fix(release): keep smoke wiring under the workflow length budget
